Move VirtualDiscFileSystem to its own file.
